Add C compile-time flag to avoid disabling ASLR
Re-exec can fail depending on how dynamic libraries are found - in addition
to being generally annoying - if readlink("/proc/self/exe") is the wrong path
to use.  The right fix is to record static space in an ELF program header to
place it where expected with no further tricks needed, and to rely on
automatic relocation of other spaces.

Compile with -DDISABLE_ASLR=0 only if needed.  Default (1 if undefined) is
to disable. (It's not "-DENABLE_ASLR" because forcing ASLR isn't a thing,
but we can forgo trying to un-enable it, hence the name chosen for this)

See https://sourceforge.net/p/sbcl/mailman/message/12418054/ for history