safepoint: Even more conditional reduction
  * Almost every time we touch all_threads_lock, it is part of a
balanced pair of lock and unlock operations.  Use the (sadly
underutilized) RUN_BODY_ONCE macro to build WITH_ALL_THREADS_LOCK,
arrange for it to compile to nothing on non-threaded builds, and
use it everywhere we can...  limited to safepoint.c, of course.

  * "I spy with my little eye something that begins with E."
"This had better not be what I..."  -- Marcus Cole and Stephen
Franklin, Babylon 5, "Racing Mars" (paraphrased)