safepoint: Pathwise substitution of thread_gc_phase() results
  * thread_gc_phase() only returns two possible results.  In a
couple of places, we save off the result, test it for one of the
possible values, and then have two blocks of code that use the
saved result instead of its known value.  This makes it harder to
find references to each phase, and harder to track what each block
is doing.

  * In these cases, where we test for an explicit result value,
substutute out the actual value wherever we use the saved result,
so that we're now using the result only the once.

  * To be fair, thread_gc_phase() was originally able to return a
third value, and it's only recently that it was determined that
said third value no longer occurs.