x86: Use packed bignum, not specialized vector, for recording fixups
This saves ~600KB in the pristine core image, but the main advantage is
having exactly one way to store code fixup locations instead of two ways.

This simplification was enabled by removal of all the different fixup fasl ops.
Those fops essentially mandated not informing GC of each fixup location until
after the fixup was intially applied, otherwise GC in the middle (if you were
so unlucky as to have the code component not referenced from the stack)
could transport the half-baked code and could mistakenly think that a fixup was
applied when it was not, and then add some random displacement to whatever
the intial value in memory was, which would then be taken as the fixup addend
when fixup application actually got around to doing that fixup. The result would
be irreparable harm to trying to deduce the correct value to install.

Fixup recording is easy now that that can't happen: whenever FIXUP-CODE-OBJECT
returns non-nil, the fixup is preserved for subsequent re-fixup by GC.
Also, the fixup packing step is not tied to sanctify-for-execution any more
because that was a confusing though necessary evil in the previous mechanism.