Micro-optimize add_new_area()
The new area array almost never fills up, so testing for overflow before
testing whether we need to record an area at all is wasteful.
In particular the root scan does not record new areas.
Additionally we do not need to check the high water mark every time
new_areas_index increases; only record it when setting index back to 0.
(We never really do anything with the high water mark other than to track
it for its own sake - maybe we can get rid of it?)