Fix, after most of nine years, "memory fault error reporting", lp#309068
  * Instead of passing the faulting address to the lisp-side error
handler via a C global variable while discarding all interrupt
context information, arrange to pass an interrupt context and the
fault address to the error handler as SAPs.  On non-x86oid targets
this is fairly easy.

  * On x86oid targets, we play bounce games similar to what
safepoints do, only slightly lighter-weight: We stash our faulting
PC and fault address on the stack and smack up the context PC to
point to a stub that contains a trap instruction for a new trap
code.  The handler for THAT recovers the faulting PC and fault
address from the stack, patches the trap context, and calls the
error handler.

  * And after all that we pull the vestiges of the old global
variable, and good riddance.

  * Net effect: the memory fault address is now interrupt and
thread safe.  And on x86 and x86-64 backtraces involving a memory
fault now have the correct top frame and can find values that are
in registers at the time of the fault.

  * Tested on x86/linux and x86-64/linux.