gencgc: Shift a bootstrap complication from C into Lisp
Genesis always writes a page table, so there is no special case
in C for not having a page table. It is a requirement to have one.
The function formerly known as gencgc_pickup_dynamic() with the
helpful comment that it acts to "Pick up the dynamic space" is gone,
and its logic is folded into gc_load_corefile_ptes().