gencgc: Combine two state variables into one that makes more sense
Or at least has a substantially better rationale and comment.

Also be a little pedantic: the allocator directly creates "regions",
not "areas", which are the coalescing of or one more adjacent regions.

There are a bunch of obsolete comments alluding to gc_alloc(),
but that's a separate matter.  The GC's maintainers of yore
were never sticklers for precision of terminology.