Fix GC lossage due to rev 0595c3f1abb1
Using WITHOUT-GCING around huge swaths of code obscured where (and why) the
code really needed to be atomic with respect to GC.

Its replacement, WITH-PINNED-OBJECTS, as used was not adequate; it left a window
in which compiling to memory could create simple-funs whose 4 boxed words were
unknown to GC because the CODE-N-ENTRIES slot was not updated yet. And that could
not be done sooner because to do so would reveal to GC the number of functions
that would be contained in the core component *before* their offsets had been
set in the code component header.  Rather than go back to WITHOUT-GCING,
rearrange so that the offsets of all simple-funs are assigned, then the number
of simple-funs, and then their boxed words. This is free of data races with GC.