x86-64: Fix bug in undoing static function linkage
Save-lisp-and-die changes all calls through fdefinitions into calls directly
to function entry points. This is completely transparent to the user,
including changing the static call back to an indirect call if encapsulating
or otherwise redefining a function. The bug was that removal of the static call
could not discern which of two different static calls was the correct one to
unpatch if two (or more) fdefns referred to the identical function.