debug-int: Rework FRAME-CATCHES and REPLACE-FRAME-CATCH-TAG
  * These are currently used only on platforms without
UNWIND-TO-FRAME-AND-CALL-VOP, but are part of the debugger
interface.  The debugger interface is a private package, but it
is also the only mechanism provided for writing a custom debugger
for SBCL, so improving it makes sense even if it might not help
with SBCL's debugger.

  * On the x86oid side, using MAKE-LISP-OBJ outside of
WITHOUT-GCING is, in the general case, unsafe.  And using it just
to pull a boxed reference relative to a SAP is unnecessary: we
have SAP-REF-LISPOBJ for that.

  * On the non-x86oid side, a catch block is not a stack frame.
It is a non-heap data structure that contains boxed values, and
should be treated as such... by using SAP-REF-LISPOBJ.

  * This resolves the gratuitous differences between x86oids and
the rest of the world.  But there are also substantive changes to
address.

  * The repeated -/+ x86oid conditionals mask that a single value
is being computed, differently in each case, due to the way that
the computation is interwoven between the two cases, and using
different variable names in certain cases.  Separate them out into
two local functions to do the computation, and pick just one to
compile using the reader conditionals.

  * Net result: fewer x86oid reader conditionals, the code is
easier to understand, and the code is safer with respect to GC
issues.