Fix handling of name and type components in DIRECTORIZE-PATHNAME
Previously, when encountering a PATHNAME with name or type components,
DIRECTORIZE-PATHNAME unparsed those via FILE-NAMESTRING and stuck that
namestring into the directory component of the result. The problem was
that the directory component is not supposed to contain
namestrings. This makes a difference when the original name or type
components contain characters that get escaped in the namestring:

  (pathname-directory (directorize-pathname (merge-pathnames "a\\*b" "/")))
  => (:ABSOLUTE "a\\*b") but should be (:ABSOLUTE "a*b")

Based on initial analyses by Richard M. Kreuter.

fixes lp#1740624