debug-int: Fix ARMoid build (I hope)
  * In theory, this is an error message for a "shouldn't happen"
case.  In practice, I'm fairly sure that I could trigger it on
quite a few platforms with a bit of jiggery-pokery (what actually
*IS* "jiggery-pokery", anyway?).

  * Anyway, ARMoids being the special little snowflakes that they
are (they're fp-and-pc-standard-save like the x86oids, but precise
like everything else, one of them uses tagged tramps like sparc,
they're the only platform that keeps the control stack pointer
outside of the register set, and they're probably special in other
ways that I just don't care to remember offhand), they have their
own little bit of code for finding a parameter for this error
message, and I didn't end up testing the build on such a platform.

  * Technically, I STILL haven't tested the build on such a
platform.  But this is a straight-up inlining of the code to
compute a value that is, necessarily, SAP= to the variable value
that is no longer lexically available after I extracted
ESCAPED-FRAME-FROM-CONTEXT from FIND-ESCAPED-FRAME.  It SHOULD fix
the build, at least.

  * I did try the crossbuild-runner this time, but I don't know if
it actually does anything that I'd notice as a failure if there is
a bug like this.

  * Also note that the value printed will be (as it always was)
different on ARMoids: the other platforms are pulling the LRA as
an unboxed value, while ARMoids are pulling it as a boxed value.
Both still print it with ~X, though.  I have no idea what that
will end up doing.

  * And also note that the actual distinction being made here is
fp-and-pc-standard-save.  And that the entire error message is
basically trying to throw any amount of data up to provide even
the slightest amount of clue as to what on earth went wrong such
that we ended up with an escaped (lisp) stack frame and no
associated (lisp) code object.  At this point, things are so badly
off-normal that... what?  This will happen if we take a signal in
call_into_lisp() on PPC after loading reg_CODE, or on any platform
(just less than half of them, I think) where we don't correctly
manage $CODE, $LRA, and the function object for call/return?  Well,
no WONDER we want to have a handle on the passing location for
the LRA.  That much, at least, is probably in-bounds, most of the
time.  The overall situation still shouldn't be happening, though.

  * Finally, this is the longest commit message that I have
written in quite some time.  Clearly, this is an important and
tricky change being made, and it requires in-depth explanation.