debug-int: More FRAME-CATCHES / REPLACE-FRAME-CATCH-TAG cleanup
  * Now that it's all SAP-REF-LISPOBJ with no architectural quirks
to contend with, shift things around so that we can use LABELS
CATCH-REF instead.

  * The next logical step in this progression is to move from
using local LABELS functions to just having a toplevel CATCH-REF
accessor.  And from there, abstract over the control flow with a
DO-CATCHES or MAP-CATCHES or something.

  * Also note that REPLACE-FRAME-CATCH-TAG distinguishes catches
based on frame and tag.  If a frame uses the same tag twice, then
both get replaced.  As this function is only ever called to
replace tag SB-C:DEBUG-CATCH-TAG, this is unlikely to be a major
issue.