Fix CMOV conversion.
BRANCH-IF and CIF may end be in different IR1 blocks (due to
cleanups), but since the BRANCH-IF is emitted from a conditional VOP
it ends up in the same IR2 block as the VOP.
MAYBE-CONVERT-ONE-CMOV tries to use (block-succ (ir2-block-block
2block)) to see where the branches lead, but the CIF is in the
following IR1 block.
Use the CIF as the node when converting BRANCH-IF, which can be used
by MAYBE-CONVERT-ONE-CMOV to find the actual block.
There's a bit of an inconsistency that a single IR2 block straddles
two IR1 blocks, but it's harmless.

Fixes lp#1741679