debug-int: Don't use GET-CONTEXT-VALUE in FRAME-SAVED-CFP-AND-LRA
  * For all FP-AND-PC-STANDARD-SAVE platforms, the OCFP and LRA
are passed in their save slots on the stack.  But their format may
vary (x86oids use SAP-STACK for the LRA, ARMoids use CONTROL-STACK,
both use CONTROL-STACK for the OCFP).

  * The simplest thing to do is to pass the buck to
SUB-ACCESS-DEBUG-VAR-SLOT, which is an old poker term referring to
the piece of buck-horn used to indicate which player was the
current dealer.  No, wait, that's not where I was going with that.
SUB-ACCESS-DEBUG-FAR-SLOT is what we're already using on
non-FP-AND-PC-STANDARD-SAVE platforms, so just tie into that
mechanism.

  * Note that this works no matter what platform is involved,
because the RETURN-PC-PASSING-OFFSET is an SC-OFFSET with the
correct SC number for the platform.  It also coincides with the
stack slot specified by LRA-SAVE-OFFSET (and the same applies for
OLD-FP-PASSING-OFFSET and OCFP-SAVE-OFFSET).

  * Why go to this trouble?  As a lead-in to trying to get rid of
GET-CONTEXT-VALUE and (SETF GET-CONTEXT-VALUE), which are dreadful.