Remove QSHOW from around debug print of pinned objects
Show large and small objects separately, and only count pins
in the generation being evacuated. There's no such thing as a pin
of a root object. Those aren't going anywhere anyway.

The printfs should really be controlled by bits of 'gencgc_verbose'
rather than the currently random choice of whether to show
based on >0, >1, or >1 plus possibly a compile-time chioce.
The only reason to compile out print statements entirely is if
they are in extremely hot code paths, and afaict, none are.

Also another fix to the ftype for GENERATION-OF, and whitespace.