debug-int: More reliable backtrace from interrupt on PPC
  * This is a new version of CODE-OBJECT-FROM-CONTEXT that handles
the holes around function call and return when the program counter
is necessarily not within the scope of the object in reg_CODE
(okay, you COULD update reg_CODE in a branch-delay slot, and some
backends DO, but then you have a window involving the branch-delay
tracking itself or... anyway, not generally safe, and so we can't
assume that reg_CODE bounds the program counter).

  * Overall, if the program counter is within the scope of some
object in a boxed register, not necessarily reg_CODE, return that
object: it is, by definition, correct.  Otherwise, one of two
situations obtains: either we're in an assembly-routine, in which
case reg_CODE is the correct thing to use, or we're in
call_into_lisp, in which case there IS no correct thing to use.

  * Currently deploying this for PPC only, because it requires a
new DEFREGSET so that the debugger knows which registers are boxed,
but the logic should be solid for all other precisely-scavenged
platforms (though it won't get a reliable backtrace if the backend
isn't actually GC-safe for interrupts around function calling,
that would take fixing the backend for GC-safety).