Pick a different name for CODE_PAGE_FLAG
FLAG implies that you can reasonably do tests like "if (allocated & x_FLAG)",
but you can't with code page flag because it's a set of bits.
More accurately, it's a point in an enumeration.  BOXED and UNBOXED are
a little of each (flag and enumerated value), and I didn't want to rename
everything, just the one that was sufficiently confusing to warrant it.