Remove per-generation alloc start pages
Every generation's collection rewinds the start page to 0, and we can't have
more than one generation being allocated to. So why keep 7 sets of state?

Also remove special case of -1 for gencgc_alloc_start_page.
This is strictly speaking a behavioral alteration, but one which only affects
gc_and_save() and not in a significant way.
The change is that on each call of gc_find_freeish_pages() when the restart page
was -1, *every* call to find free pages searched upward from the specified page
and had a chance of finding gaps left due to earlier allocations. Gaps occur
when free space in a region is inadequate and a new region is started at the
next available page boundary. Those gaps would be usable under the old scheme.

It was quite an ugly special case, because it is almost equivalently handled
by the new code that makes no special exception, but still arranges to move
the whole dynamic space to a page range higher than last_free_page prior to
moving it all back downward again. Filling in of gaps during the penultimate GC
won't happen, because there is no explicit rewind, and so it might consume
a slightly larger number of pages than it used to.  The final GC is unaffected
by this change, and if it doesn't fit in its allotted range, that's bad,
but not a regression. (It was never guaranteed - see the new comments)