Move operation out of loop in free_oldspace()
The loop termination condition was based on seeing a page not in from_space,
but the loop body seemed not to understand that the page's generation was
a loop invariant. And flip the order of tests in the termination condition
because page_bytes_used is a worse discriminator than generation.