Increase aggressiveness of tree shaker when saving initial core
Prior to saving, we unintern all symbols, pointing weakly to them,
then collect garbage (starting only at static roots and stacks),
and re-intern the non-broken pointers.
Any symbol with a reference chain from a known root is always retained.
A symbol having no incoming reference except for a package is dealt
with in one of the following manners dependent on its home package:
 1. require preserving all external symbols
    (with some new exports in package-data-list to make this work)
 2. preserve only specific (ad-hoc) internal symbols
 2. preserve externals plus some specific symbols
 4. preserve any symbol with any interesting properties at all-
    function binding, value, globaldb data, etc.

The result is more than a 4% core size reduction with x86-64.

It is likely that this will discover occurences of add-on libraries
poking around with SBCL-internal symbols that disappeared.
This is as intended even if means having to unbreak some things
by fixing 3rd-party code or adding more symbol preservation exceptions.