x86-64: avoid register load to call asm routines
In terms of performance this is neutral, although the CPU would probably
prefer not to see data acccess in the same cache line as instruction fetch.
So the jump vectors should really precede all the code. That's hard to fix
though since we don't know how many routines to leave space for.