Fix cross-thread pollution of *COMPILER-SSET-COUNTER*
*COMPILER-SSET-COUNTER* is freely referenced by at least these callers:
  SB-PCL::WALK-FORM
  SB-INTERPRETER::LEXENV-FROM-ENV [and possibly SB-EVAL]
  SB-CLTL2::AUGMENT-ENVIRONMENT

This was, if nothing else, a source of a data race if multiple threads
use the PCL walker etc. Of the possible ways to resolve the problem
of failing to bind *COMPILER-SSET-COUNTER*, the following were rejected:

* Always explicitly pass :NUMBER NIL when making instances of LAMBDA-VAR,
  GLOBAL-VAR etc. in the PCL walker etc. (i.e. tediously fix all callers)

* Use a single global counter, never rebind, and always atomically increment.
  This would allow the non-compiler producers of LAMBDA-VAR to create legal
  ones in as much as they can be SSET members.
  But non-binding is ugly and results in wraparound arbitrarily somewhere.

* Aways bind the variable if referenced. This works for LEXENV-FROM-ENV and
  WALK-FORM in that they can scope the binding, and so instances created by
  by those could be sequntially numbered. It doesn't work so well for
  AUGMENT-ENVIRONMENT though, and requires touching all callers.

Ultimately, it just seems best to avoid numbering any instance that was not
created within the compiler. So after this change, producers other than the
compiler intentially do not make instances that can be added to an SSET.

It would have been nice if this actually fixed threadsafety, but, alas ... no.
The counter is a mere hash, and strict uniqueness was never a requirement.