Compiler threadsafety fix - bind *ADJUSTABLE-VECTORS*
The cost of one vector is immeasurably small relative to the cost
of making hundreds to thousands of TNs per code component.

Remove define-thread-local on *compiler-sset-counter*. It is bound in the
compiler and never assigned unless it has other than the global default value.
Prior change affecting it was in part noise (and in part relevant).