Implement a new syntax element to #!+ feature reader
  "#!+(VOP-TRANSLATES X)" produces a form only if a vop translates function X.
  "#!+(VOP-EXISTS X)" produces a form only if a vop is named X.

This was initially a workaround for the fact that CLISP is unable to read
a "#." producing 0 values, but it seems like it should be preferred to using
a bunch of different features describing presence of backend-specific vops.
Relying on the vop definitions themselves as the source of truth for what
vops exist is simpler than keeping make-config.sh in sync.

Potentially all of these (and more) could go away:
* ash-right-vops
* compare-and-swap-vops
* complex-float-vops
* float-eql-vops
* integer-eql-vop
* memory-barrier-vops
* multiply-high-vops
* raw-instance-init-vops
* symbol-info-vops
* unwind-to-frame-and-call-vop
* unbind-n-vop

The difficulty is that if the file needing a feature test is placed in the
build order prior to compiler/target/*.lisp then the feature queries might
report the wrong answer, and it would be tough to know that.