Test for case that ELF + heap relocation can't handle, and abort.
ELF isn't used by default, and the only pressing need for relocation is
the memory sanitizer, which conflicts with dynamic but not fixedobj space,
so this isn't removing any capability in the standard build.

Croaking explicitly in C rather than somewhere randomly in Lisp reminds
me what remains to be done rather than having to track it down every time.