Change lots of DEFGLOBALs to DEFINE-LOAD-TIME-GLOBAL
As mentioned in several other changes, it is unfortunate that DEFGLOBAL
implies compile-time evaluation since it is always possible to opt-in
to a compile-time effect via eval-when. But that ship has sailed,
and we have spell this out using two different macros.

So this change touches a subset of DEFGLOBALs seen in make-host-2 but not
make-host-1.  If a given DEFGLOBAL is seen during both passes, changing it
to DEFINE-LOAD-TIME-GLOBAL is no different, because the symbol already
satisfies BOUNDP.  But if only seen in make-host-2, this change explicitly
avoids reliance on the compile-time value.