x86-64: Allow immediate SC as 'value' in cell setters
This deals with:
 * all object types that goes through CELL-SET, such as CONS
 * simple-vector
 * instance

The restriction to (signed-byte 32) for the immediate is removed.
In the case where the result of the setter is used, and the immediate operand
would have been loaded on behalf of the vop, and moved into a different TN
for the result, we shorten:
  MOV RAX, $some_constant
  MOV [RDI+disp], RAX
  MOV RDX, RAX ; move result
to:
  MOV RDX, $some_constant ; move result
  MOV [RDI+disp], RDX
by using the result TN to load the immediate operand of the store.