Limit CLOSURE-INIT value SCs to {DESCRIPTOR,ANY}-REG
Other SCs crept in due to change c66b6b67c3 which removed SLOT-SET.
Confusingly we had both SET-SLOT and SLOT-SET, the latter of which accepted
an IMMEDIATE value, and perhaps NULL or ZERO depending on the backend.
The acceptable SCs from SLOT-SET were just blindly copied to CLOSURE-INIT.

Semantically, immediate SCs can't be useful. If a value were initially known
but subject to assignment, then the closure would be over a value cell, not
the initial value. And if the value were known and never assigned,
then it wouldn't be a closed-over value - it would be a literal constant.

The flip side of this is that VALUE-CELL-SET ought to accept immediate
operands, but (at least on x86-64), the compiler doesn't use it as such. e.g.:

* (disassemble '(lambda () (let ((x 1)) (lambda () (incf x)))))
 ... some stuff
 ; 48:       B802000000       MOV EAX, 2
 [11 lines omitted]
 ; 79: L0:   488942F9         MOV [RDX-7], RAX

which shows it initializing a value-cell from a preloaded register.