x86-64: Unpessimize LOGTEST
Integer comparison vops lost the ability to emit mem + immediate mode
after change 39117fb4ca which was supposed to improve recognition of positive
64-bit constants bitwise equal to a negative imm32 operand. Removal of the
complicated :LOAD-IF to accept small tagged numbers caused a regression.

After thinking about this for a while, I realized that :load-if expressions
for x86-64 willl forever be incorrect, a mess of spaghetti, or unmaintainable.
(Choose any three.) Instead we should accept register and stack operands
in all arithmetic vops, letting the compiler think that the CPU can directly
deal with it. It can't always, but since R11 is permanently reserved for lots
of other reasons anyway, we can load/store through R11 just-in-time if needed.

Rather than attempt a massive change all at once, the current change only
deals with LOGTEST and adds some generally useful utilities. Additionally
this change adds an optimization to use high byte legacy registers,
which is an orthogonal optimization, but affecting the same piece of code.