Append *ELSEWHERE* to *CODE-SEGMENT* after each asm routine
Makes each routine's code logically self-contained and, on x86,
shortens the encoding of the JMP to elsewhere and back.

We never cared about alignment of elsewhere (since each routine's
elsewhere just got piled on to the end of the preceding one)
but if we wanted to, it would be simple enough to insert alignment.