Close the current thread's allocation region in MAP-ALLOCATED-OBJECTS
to enable visiting of objects recently consed by the calling thread.
It's OK to do because the C code acquires free_pages_lock.

As to seeing other threads' consing, the mapping operation always
understates the heap contents if multiple threads are actively consing,
or even idle but in a state of having consed anything since last GC.
This patch doesn't make anything worse or better in that regard.