initial steps towards METHOD-COMBINATION rethinking
FIND-METHOD-COMBINATION, we think, would have been better named
COMPUTE-EFFECTIVE-METHOD-CONSTRUCTOR-USING-GENERIC-FUNCTION.  As such,
its primary job is to allow for different ways of choosing a method
combination, based on the class of the generic function that it will
be attached to.  This seems like speculative generality to say the
least, but it does mean that for standard generic functions we should
just do what we do.

Doing what we do is currently done by a confusing mechanism: computing
methods to add to FIND-METHOD-COMBINATION which basically do the same
thing (with slight differences based on whether it's a short or long
form method combination).  I think this complexity is confusing and,
based on this patch, unnecessary; just record the name of a method
combination type in a hash table instead, keying to some
information (a constructor function and a source location).

While we're at it, we might as well implement a cache for method
combination objects, essentially turning each fully resolved method
combination object (i.e. with both method-combination-define-time and
generic-function-define-time options stated) into a singleton.  This
will be useful when we want to have links from method combination
objects to the generic functions that use them, to support method
combination redefinition.

This patch should change no user-visible behaviour at all.  In
particular, all the bugs around redefinitinos of method combinations
still exist.

(Committing and pushing from the airport like a pro.)