Remove union per_thread_data
Cast 'thread' to a (lispobj*) and be done with it. Taking the address of a
union, which is essentially what we did everywhere, is not a useful pattern.

Or to put it another way, the purpose of a union is to avoid casting its
memory contents each time accessed. If you can't avoid casts anyway,
and the code looks worse for it, then a union is possibly the wrong choice.