Change heuristic from package-name to package-locked-p
The decision to perform an optimization based on package name is horrible.

As an aside- This is the same problem we'll have if we implement external
entries that receive unboxed args, etc, while also allowing incompatible
redefinitions. We will probably want to apply such optimizations only when
signatures are promised, or else we'd have to (weakly) track caller-callee
relations and be in a position to actually recompile callers.
The latter is neither feasible nor necessary for production code.