Simplify the calling of assembler instruction hooks
Receive all operands as a single &REST list in each instruction emitter,
passing the whole list through to INSTRUCTION-HOOKS instead of deconstructing
and reconstructing the args through a ridiculously complicated M-V-CALL form.

Now that the compiler is responsible for argument parsing, it can style-warn
about unused instruction operands; indeed there were some, which is weird.

And while the complicated code that was removed might have in theory avoided
consing some lists, it's not clear that it actually achieved that aim,
and it's certainly irrelevant in the scheme of things.