Store linkage table addresses as offsets from the table base
If we want to make linkage space relocatable (not to suggest that it's
easy or even possible), then table-relative values will make more sense.

And for detecting overflow, check if the entry being written ends at or
below the space end, not whether it starts strictly below the end,
in case the table size is not an exact multiple of the entry size.