Avoid adjustment of boxed word count in code allocators
Give allocate-code-object the number of constants needed, and adjust there
for the fixed overhed and alignment. Do not further adjust in C or assembly.

This potentially allows squeezing out 2 words if there is a padding word
following the boxed constants and also following the unboxed constants.
For the time being, the alignment constraints remain unchanged.

Only tested lightly on cheneygc. (arm64 probably was always wrong
because it assumed, like others, that WORD-SHIFT = N-FIXNUM-TAG-bits)