Make codegen store a table of simple-fun offsets in code object
Rather than repeating the same thing in make-core-component and the loader
and genesis, build a table in only one place, essentially at assembly time.

The table of offsets built by codegen is stored after the simple-funs,
not with unboxed constants. Calculation of the number of instruction bytes
in a code object should generally be done with %CODE-TEXT-SIZE,
not %CODE-CODE-SIZE, which includes trailing raw data.

Temporarily, both new and old representations are present.
The older one is going to be removed in a subsequent change.