x86-64: Remove :CODE-OBJECT and :CLOSURE fixups
This was a misuse of fixups. :CODE-OBJECT is meaningful only in 32-bit mode
which lacks RIP-relative addressing. In that case, the fixed-up value will
eventually be based on the actual address of the newly allocated code object.

But in x86-64, these are relative offsets fully resolved by the assembler,
and never dumped as fixups. Not only was it misleading, the way EMIT-EA handled
them was fubar: detect that we didn't really desire a fixup, then recursively
invoke EMIT-EA, then figure out that we still didn't want a fixup, then invoke
EMIT-LABEL-RIP.  Instead just use an EA that expresses RIP-relative mode.