Unbreak 32-bit darwin (broken by "Delete get_spinlock")
Because we don't adhere to "Include What You Use", it is possible to induce
subtle bugs that doesn't even show up at compile-time. In this case, absent
the declaration of lose() from an indirect includer of x86-arch, there was a
void function that supposedly returned, which would not be a problem per se,
except that undefined behavior lets the C compiler do anything it wants.