x86-64: Relax constraint on boxed header word count
Remove two padding words from code objects in the following scenario:

  { boxed word }    { raw data }    { simple-fun-header }
  { ...        } +  { ...      } +  { etc.              }
  { padding    }    { padding  }
  --------------    ------------
  even length    +  even length  = even length

by changing it to odd + odd = even, provided that first word of raw data needs
fewer than 4 bits of alignment, as is true of anything smaller than octword.

And if the boxed count is odd but no unboxed constants exist, it may as well
be the case that GC regards the mandatory padding as belonging to unboxed bytes.

This change should work on all backends but has not been tested.
(EMIT-HEADER-DATA needs to be altered, or just put into a common place
since it's pretty much identical across all compiler/*/insts.lisp files)