gencgc: Enforce code/data separation invariant in genesis
Genesis will never mix code and non-code on the same GC page, just as
the C allocator won't. Producing the initial page table is also more
understandable. The removed logic for computing scan-start was quite
wrong and did not agree at all with the new logic, which is right.

So now instead of CODE_PAGE_TYPE meaning that a page MAY contain code,
it will mean that the page MUST contain code, and nothing but code.
This allows some further improvements to the garbage collector.