Remove SC_GEN_CK - it can not work
If there are multiple threads involved - which is most likely when you want
maximal GC debugging (if you need it at all) - then this code was technically
wrong, i.e. broken. To make it right, the wp_cleared bit would need to be
restored after the check. Better off to do a pre-verify of the heap for any
pages with the WP bit on containing old -> young pointers.
Also fwiw, the comment about weak pointers was incorrect.