Stop binding *default-pathname-defaults* in SB-COVER:REPORT
The code had a telltale sign of using two different techniques for choosing the
location for per-file reports: (1) mirroring the original source tree in a
similarly shaped output tree of HTML files; (2) using a single output dir
and naming each output file as a hash of the original namestring.
My hunch is that the former never worked, as anyone who has ever wrestled
with Lisp's obsession with premature merging of pathnames will have discovered.

Anyway, it's wrong to rebind *D-P-DEFAULTS* because if the source namestrings
are stored as names relative to the current directory (which is the right way
as per all modern tooling), but not the Lisp way, or at least not our default
way, then it doesn't work to try to merge those relative names with whatever
random location you asked for as your output directory.