Build features as special host features
If you wanted to control part of the build process, one way was using
regular target features. An example of this bad use case for target
features was sb-after-xc-core.
This introduces the --build syntax, that adds features into the host's
features instead.
It provides retro-compatibility for the old usage of sb-after-xc-core

committer: Douglas Katzman <dougk@google.com>