Make editcore not croak in verify_heap()
It was confusing to create cons cells in the local memory whose parts
are pointers into "remote" memory using untranslated bits, i.e. the bits
as they are in the foreign core. For example, don't cons this:

  (#x100010023f . 0) ; <-- the cons is in the local core @ #x1000a00007
     ^--- the car points to the "remote" core, but it looks local

In most cases, that cons would "see junk" unless by sheer luck, the
remote pointer's bits were valid enough for verify_heap to accept.
So wrap remote pointers in a struct with 1 slot holding opaque bits.