Make --build-cons-profiling work as intended
The distinction between target vs build features is that you're supposed to
be able to configure the target and run make-host-1, but choose build options
for varying invocations of make-host-2 without changing sb!xc:*features*.
The ordinary build scripts don't make use of this ability, but customized
scripts can, to obtain (in parallel!) different cold cores from a single
build of the cross-compiler. And if nothing else, this removes more reader
conditionals than it adds, which is strictly an improvement.